# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Match": {
        "properties": {
          "court_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Court Id"
          },
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "custom_duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Duration Minutes"
          },
          "custom_margin_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Margin Minutes"
          },
          "duration_minutes": {
            "title": "Duration Minutes",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "margin_minutes": {
            "title": "Margin Minutes",
            "type": "integer"
          },
          "position_in_schedule": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position In Schedule"
          },
          "round_id": {
            "title": "Round Id",
            "type": "integer"
          },
          "stage_item_input1": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageItemInputTentative"
              },
              {
                "$ref": "#/components/schemas/StageItemInputFinal"
              },
              {
                "$ref": "#/components/schemas/StageItemInputEmpty"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input1"
          },
          "stage_item_input1_conflict": {
            "title": "Stage Item Input1 Conflict",
            "type": "boolean"
          },
          "stage_item_input1_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input1 Id"
          },
          "stage_item_input1_score": {
            "title": "Stage Item Input1 Score",
            "type": "integer"
          },
          "stage_item_input1_winner_from_match_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input1 Winner From Match Id"
          },
          "stage_item_input2": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageItemInputTentative"
              },
              {
                "$ref": "#/components/schemas/StageItemInputFinal"
              },
              {
                "$ref": "#/components/schemas/StageItemInputEmpty"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input2"
          },
          "stage_item_input2_conflict": {
            "title": "Stage Item Input2 Conflict",
            "type": "boolean"
          },
          "stage_item_input2_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input2 Id"
          },
          "stage_item_input2_score": {
            "title": "Stage Item Input2 Score",
            "type": "integer"
          },
          "stage_item_input2_winner_from_match_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input2 Winner From Match Id"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          }
        },
        "required": [
          "created",
          "start_time",
          "duration_minutes",
          "margin_minutes",
          "custom_duration_minutes",
          "custom_margin_minutes",
          "position_in_schedule",
          "round_id",
          "stage_item_input1_score",
          "stage_item_input2_score",
          "court_id",
          "stage_item_input1_conflict",
          "stage_item_input2_conflict",
          "stage_item_input1_id",
          "stage_item_input2_id",
          "stage_item_input1_winner_from_match_id",
          "stage_item_input2_winner_from_match_id",
          "id",
          "stage_item_input1",
          "stage_item_input2"
        ],
        "title": "Match",
        "type": "object"
      },
      "MatchBody": {
        "properties": {
          "court_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Court Id"
          },
          "custom_duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Duration Minutes"
          },
          "custom_margin_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Margin Minutes"
          },
          "round_id": {
            "title": "Round Id",
            "type": "integer"
          },
          "sets": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MatchSetBody"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sets"
          },
          "stage_item_input1_score": {
            "default": 0,
            "title": "Stage Item Input1 Score",
            "type": "integer"
          },
          "stage_item_input2_score": {
            "default": 0,
            "title": "Stage Item Input2 Score",
            "type": "integer"
          }
        },
        "required": [
          "round_id",
          "stage_item_input1_score",
          "stage_item_input2_score",
          "court_id",
          "custom_duration_minutes",
          "custom_margin_minutes",
          "sets"
        ],
        "title": "MatchBody",
        "type": "object"
      },
      "MatchCreateBodyFrontend": {
        "properties": {
          "court_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Court Id"
          },
          "round_id": {
            "title": "Round Id",
            "type": "integer"
          },
          "stage_item_input1_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input1 Id"
          },
          "stage_item_input1_winner_from_match_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input1 Winner From Match Id"
          },
          "stage_item_input2_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input2 Id"
          },
          "stage_item_input2_winner_from_match_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Item Input2 Winner From Match Id"
          }
        },
        "required": [
          "round_id",
          "court_id",
          "stage_item_input1_id",
          "stage_item_input2_id",
          "stage_item_input1_winner_from_match_id",
          "stage_item_input2_winner_from_match_id"
        ],
        "title": "MatchCreateBodyFrontend",
        "type": "object"
      },
      "MatchRescheduleBody": {
        "properties": {
          "new_court_id": {
            "title": "New Court Id",
            "type": "integer"
          },
          "new_position": {
            "title": "New Position",
            "type": "integer"
          },
          "old_court_id": {
            "title": "Old Court Id",
            "type": "integer"
          },
          "old_position": {
            "title": "Old Position",
            "type": "integer"
          }
        },
        "required": [
          "old_court_id",
          "old_position",
          "new_court_id",
          "new_position"
        ],
        "title": "MatchRescheduleBody",
        "type": "object"
      },
...
      "MatchSetBody": {
        "properties": {
          "score1": {
            "default": 0,
            "minimum": 0.0,
            "title": "Score1",
            "type": "integer"
          },
          "score2": {
            "default": 0,
            "minimum": 0.0,
            "title": "Score2",
            "type": "integer"
          },
          "set_number": {
            "minimum": 1.0,
            "title": "Set Number",
            "type": "integer"
          }
        },
        "required": [
          "set_number",
          "score1",
          "score2"
        ],
        "title": "MatchSetBody",
        "type": "object"
      },
...
      "SingleMatchResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Match"
          }
        },
        "required": [
          "data"
        ],
        "title": "SingleMatchResponse",
        "type": "object"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "SuccessResponse",
        "type": "object"
      },
      "SuggestedMatch": {
        "properties": {
          "elo_diff": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Elo Diff",
            "type": "string"
          },
          "is_recommended": {
            "title": "Is Recommended",
            "type": "boolean"
          },
          "player_behind_schedule_count": {
            "title": "Player Behind Schedule Count",
            "type": "integer"
          },
          "stage_item_input1": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageItemInputTentative"
              },
              {
                "$ref": "#/components/schemas/StageItemInputFinal"
              },
              {
                "$ref": "#/components/schemas/StageItemInputEmpty"
              }
            ],
            "title": "Stage Item Input1"
          },
          "stage_item_input2": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageItemInputTentative"
              },
              {
                "$ref": "#/components/schemas/StageItemInputFinal"
              },
              {
                "$ref": "#/components/schemas/StageItemInputEmpty"
              }
            ],
            "title": "Stage Item Input2"
          },
          "swiss_diff": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Swiss Diff",
            "type": "string"
          },
          "times_played_sum": {
            "title": "Times Played Sum",
            "type": "integer"
          }
        },
        "required": [
          "stage_item_input1",
          "stage_item_input2",
          "elo_diff",
          "swiss_diff",
          "is_recommended",
          "times_played_sum",
          "player_behind_schedule_count"
        ],
        "title": "SuggestedMatch",
        "type": "object"
      },
...
      "UpcomingMatchesResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SuggestedMatch"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "UpcomingMatchesResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\n### Description\nThis API allows you to do everything the frontend of [Bracket](https://github.com/evroon/bracket)\nallows you to do (the frontend uses this API as well).\n\nFore more information, see the [documentation](https://docs.bracketapp.nl).\n\n### Table of Contents\n*(links only work for [ReDoc](https://api.bracketapp.nl/redoc), not for Swagger UI)*\n\n- [Auth](#tag/Auth)\n\n- [Clubs](#tag/Clubs)\n\n- [Courts](#tag/Courts)\n\n- [Internals](#tag/Internals)\n\n- [Matches](#tag/Matches)\n\n- [Players](#tag/Players)\n\n- [Rankings](#tag/Rankings)\n\n- [Rounds](#tag/Rounds)\n\n- [Sport Configs](#tag/Sport-Configs)\n\n- [Stage Items](#tag/Stage-Items)\n\n- [Stage Item Inputs](#tag/Stage-Item-Inputs)\n\n- [Stages](#tag/Stages)\n\n- [Teams](#tag/Teams)\n\n- [Tournaments](#tag/Tournaments)\n\n- [Users](#tag/Users)\n\n### Links\nGitHub: <https://github.com/evroon/bracket>\n\nDocs: <https://docs.bracketapp.nl>\n\nDemo: <https://www.bracketapp.nl/demo>\n\nAPI docs (Redoc): <https://api.bracketapp.nl/redoc>\n\nAPI docs (Swagger UI): <https://api.bracketapp.nl/docs>\n",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "summary": "API for Bracket, an open source tournament system.",
    "title": "Bracket API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tournaments/{tournament_id}/matches": {
      "post": {
        "operationId": "create_match_tournaments__tournament_id__matches_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchCreateBodyFrontend"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleMatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Match",
        "tags": [
          "Matches"
        ]
      }
    },
    "/tournaments/{tournament_id}/matches/{match_id}": {
      "delete": {
        "operationId": "delete_match_tournaments__tournament_id__matches__match_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Match",
        "tags": [
          "Matches"
        ]
      },
      "put": {
        "operationId": "update_match_by_id_tournaments__tournament_id__matches__match_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Match By Id",
        "tags": [
          "Matches"
        ]
      }
    },
    "/tournaments/{tournament_id}/matches/{match_id}/reschedule": {
      "post": {
        "operationId": "reschedule_match_tournaments__tournament_id__matches__match_id__reschedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRescheduleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Reschedule Match",
        "tags": [
          "Matches"
        ]
      }
    },
...
    "/tournaments/{tournament_id}/schedule_matches": {
      "post": {
        "operationId": "schedule_matches_tournaments__tournament_id__schedule_matches_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Schedule Matches",
        "tags": [
          "Matches"
        ]
      }
    },
...
    "/tournaments/{tournament_id}/stage_items/{stage_item_id}/upcoming_matches": {
      "get": {
        "operationId": "get_matches_to_schedule_tournaments__tournament_id__stage_items__stage_item_id__upcoming_matches_get",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "stage_item_id",
            "required": true,
            "schema": {
              "title": "Stage Item Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "elo_diff_threshold",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Elo Diff Threshold",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "iterations",
            "required": false,
            "schema": {
              "default": 2000,
              "title": "Iterations",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "only_recommended",
            "required": false,
            "schema": {
              "default": false,
              "title": "Only Recommended",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpcomingMatchesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Matches To Schedule",
        "tags": [
          "Matches"
        ]
      }
    },
...
  }
}